toolkit/components/url-classifier/tests/mochitest/test_classifier.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test the URI Classifier</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

var Cc = SpecialPowers.Cc;
var Ci = SpecialPowers.Ci;
var firstLoad = true;

// Add some URLs to the malware database.
var testData = "malware.example.com/";
var testUpdate =
  "n:1000\ni:test-malware-simple\nad:1\n" +
  "a:524:32:" + testData.length + "\n" +
  testData;

var dbService = Cc["@mozilla.org/url-classifier/dbservice;1"]
                .getService(Ci.nsIUrlClassifierDBService);

function loadTestFrame() {
  document.getElementById("testFrame").src = "classifierFrame.html";
}

function doUpdate(update) {
  var listener = {
    QueryInterface: function(iid)
    {
      if (iid.equals(Ci.nsISupports) ||
          iid.equals(Ci.nsIUrlClassifierUpdateObserver))
        return this;

      throw Cr.NS_ERROR_NO_INTERFACE;
    },
    updateUrlRequested: function(url) { },
    streamFinished: function(status) { },
    updateError: function(errorCode) {
      ok(false, "Couldn't update classifier.");
      // Abort test.
      SimpleTest.finish();
    },
    updateSuccess: function(requestedTimeout) {
      SpecialPowers.pushPrefEnv(
        {"set" : [["browser.safebrowsing.malware.enabled", true]]},
        loadTestFrame);
    }
  };

  dbService.beginUpdate(listener, "test-malware-simple", "");
  dbService.beginStream("", "");
  dbService.updateStream(update);
  dbService.finishStream();
  dbService.finishUpdate();
}

SpecialPowers.pushPrefEnv(
  {"set" : [["urlclassifier.malwareTable", "test-malware-simple"],
            ["urlclassifier.phishTable", "test-phish-simple"]]},
  function() { doUpdate(testUpdate); });

// Expected finish() call is in "classifierFrame.html".
SimpleTest.waitForExplicitFinish();

</script>

</pre>
<iframe id="testFrame" onload=""></iframe>
</body>
</html>
